+++ server/e2e_tests/playwright.config.ts
@@ -1,0 +1,7 @@
+import { PlaywrightTestConfig } from '@playwright/test';
+
+const config: PlaywrightTestConfig = {
+  use: { screenshot: 'only-on-failure' },
+  fullyParallel: true,
+};
+export default config;