+++ server/e2e_tests/test-config.ts
@@ -1,0 +1,11 @@
+import { TestConfig } from "./e2e-generated.spec";
+const demoFileName = 'logo.svg';
+
+export const testConfig: TestConfig = {
+  demoFileName,
+  demoFile: `../../${demoFileName}`,
+  demoInviteName: 'document demo',
+  serverUrl: 'http://localhost:9883',
+  frontEndUrl: 'http://localhost:9883',
+  initialTest: true,
+};