+++ CHANGELOG.md
@@ -11,6 +11,7 @@ By far most changes relate to `atomic-server`, so if not specified, assume the c
 - Fix default config directory, set it again to `~/.config/atomic`. This accidentally was `~` since v0.32.0.
 - Fix flaky query test #468
 - Don't subscribe to external resources #470
+- Improve frequency search indexing #473
 
 ## [v0.32.2] - 2022-06-20